backend/requirements.txt
# Core FastAPI and ASGI server
fastapi
uvicorn

# Database (PostgreSQL)
sqlalchemy
psycopg2-binary

# Google Cloud BigQuery integration
google-cloud-bigquery

# Data validation and settings management
pydantic
python-dotenv

# Database migrations
alembic

# LLM Integration (placeholder for now, specific library depends on LLM chosen)
# For example, if using OpenAI:
# openai
# For Google Cloud Vertex AI:
# google-cloud-aiplatform
